requirements.txt
# E-commerce CLIP Fine-tuning Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0

# CLIP
git+https://github.com/openai/CLIP.git

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Data handling
pandas>=2.0.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Development tools (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# CUDA support (if using GPU)
# Note: Install appropriate CUDA version for your system
# torch with CUDA: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
